requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
docx-mailmerge==0.4.0
google-api-python-client==2.178.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
idna==3.10
lxml==6.0.0
PyPDF2==3.0.1
PyQt6==6.9.1
requests==2.32.4
soupsieve==2.7
urllib3==2.5.0
youtube-transcript-api==1.2.2
